Changes distance to double

+++ core/domain/event.go
@@ -6,6 +6,6 @@ type Event struct {
 	Id       int       `json:"id"`
 	Name     string    `json:"name"`
 	Location string    `json:"location"`
-	Distance int32     `json:"distance"`
+	Distance float64   `json:"distance"`
 	Date     time.Time `json:"date"`
 }
